requirements.txt
Babel==2.7.0
dj-database-url==0.5.0
Django==2.2.4
django-bulma==0.5.6
django-heroku==0.3.1
django-money==0.15
django-phonenumber-field==3.0.1
phonenumbers==8.10.16
Pillow==6.1.0
psycopg2==2.8.3
py-moneyed==0.8.0
python-decouple==3.1
pytz==2019.2
sqlparse==0.3.0
whitenoise==4.1.3
yapf==0.28.0
